+++ internal/ipfs/wrap.go
@@ -3,10 +3,12 @@ package ipfs
 import (
 	"fmt"
 	"regexp"
+	"strings"
 )
 
 const ipfsLinkTemplate = "https://ipfs.io/ipfs/%s"
 const ipfsDAOImageLinkTemplate = "https://cdn.stamp.fyi/space/%s?s=90"
+const ipfsResolveURL = "https://gateway.4everland.link/ipfs/"
 
 var linkRE = regexp.MustCompile(`^ipfs://([a-zA-Z0-9]+)$`)
 
@@ -21,4 +23,8 @@ func WrapLink(link string) string {
 
 func WrapDAOImageLink(ensName string) string {
 	return fmt.Sprintf(ipfsDAOImageLinkTemplate, ensName)
+}
+
+func ReplaceLinksInText(text string) string {
+	return strings.ReplaceAll(text, "ipfs://", ipfsResolveURL)
 }